Added decimals and colors

+++ app/src/main/java/wayne/com/javadontliekotlin/data/db/Game.kt
@@ -9,6 +9,6 @@ import kotlin.collections.ArrayList
 @Entity(tableName = "Games")
 class Game(@PrimaryKey(autoGenerate = true) var id: Long? = null,
            @ColumnInfo(name = "gameName") var gameName: String = "",
-           @ColumnInfo(name = "amount") var amount: Int = 0,
+           @ColumnInfo(name = "amount") var amount: Double = 0.0,
            @ColumnInfo(name = "log") var log: ArrayList<String> = ArrayList()
 )
